# silvera-implementation/scientific-paper-publisher/output/User/openapi.json
{"openapi": "3.0.0", "info": {"version": "0.0.1b", "title": "User", "description": ""}, "servers": [{"url": "http://localhost:9095/api/user"}], "paths": {"/user": {"post": {"description": "Creates a new user.", "operationId": "createUser", "requestBody": {"description": "User to add.", "required": true, "content": {"application/json": {"schema": {"$ref": "#components/schemas/User"}}}}, "responses": {"200": {"$ref": "#components/responses/UserOK"}}}}, "/user/id": {"put": {"description": "Update User object.", "operationId": "updateUser", "parameters": [{"name": "id", "in": "path", "description": "ID of user to update", "required": true, "schema": {"type": "string"}}], "requestBody": {"$ref": "#components/requestBodies/UserBody"}, "responses": {"200": {"$ref": "#components/responses/UserOK"}}}, "get": {"description": "Read User object.", "operationId": "readUser", "parameters": [{"name": "id", "in": "path", "description": "ID of user to update", "required": true, "schema": {"type": "string"}}], "requestBody": {"$ref": "#components/requestBodies/UserBody"}, "responses": {"200": {"$ref": "#components/responses/UserOK"}}}, "delete": {"description": "Delete User object.", "operationId": "deleteUser", "parameters": [{"name": "id", "in": "path", "description": "ID of user to update", "required": true, "schema": {"type": "string"}}], "requestBody": {"$ref": "#components/requestBodies/UserBody"}, "responses": {"200": {"$ref": "#components/responses/UserOK"}}}}, "/login": {"post": {"description": "", "operationId": "logIn", "parameters": [], "requestBody": {"description": "JSON representation object", "required": true, "content": {"application/json": {"schema": {"$ref": "#components/schemas/LoginInfo"}}}}, "responses": {"200": {"description": "JSON representation of string object", "content": {"application/json": {"schema": {"type": "string"}}}}}}}, "/isloggedin/{username}": {"get": {"description": "", "operationId": "isLoggedIn", "parameters": [{"name": "username", "in": "path", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "JSON representation of boolean object", "content": {"application/json": {"schema": {"type": "boolean"}}}}}}}, "/getname/{username}": {"get": {"description": "", "operationId": "getName", "parameters": [{"name": "username", "in": "path", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "JSON representation of string object", "content": {"application/json": {"schema": {"type": "string"}}}}}}}}, "components": {"schemas": {"User": {"type": "object", "properties": {"username": {"type": "string"}, "password": {"type": "string"}, "firstName": {"type": "string"}, "lastName": {"type": "string"}, "loggedIn": {"type": "boolean"}, "email": {"type": "string"}}, "required": ["username", "password", "firstName", "lastName"]}, "LoginInfo": {"type": "object", "properties": {"username": {"type": "string"}, "password": {"type": "string"}}, "required": ["username", "password"]}}, "requestBodies": {"UserBody": {"description": "JSON representation of User object", "required": true, "content": {"application/json": {"schema": {"$ref": "#components/schemas/User"}}}}, "LoginInfoBody": {"description": "JSON representation of LoginInfo object", "required": true, "content": {"application/json": {"schema": {"$ref": "#components/schemas/LoginInfo"}}}}}, "responses": {"UserOK": {"description": "JSON representation of User object", "content": {"application/json": {"schema": {"$ref": "#components/schemas/User"}}}}, "LoginInfoOK": {"description": "JSON representation of LoginInfo object", "content": {"application/json": {"schema": {"$ref": "#components/schemas/LoginInfo"}}}}}}}