# silvera-implementation/scientific-paper-publisher/output/User/src/main/java/com/silvera/User/domain/model/User.java
/**
    THIS IS GENERATED CODE AND SHOULD NOT BE CHANGED MANUALLY!!!

    Generated by: silvera
    Date: 2022-05-19 13:06:03
*/

package com.silvera.User.domain.model;

import org.springframework.data.annotation.Id;
import javax.validation.constraints.*;
public class User {

    
    
    
    @Id
    
    @NotBlank(message="Field 'username' cannot be empty!")
    private java.lang.String username;
    
    
    @NotBlank(message="Field 'password' cannot be empty!")
    private java.lang.String password;
    
    
    @NotBlank(message="Field 'firstName' cannot be empty!")
    private java.lang.String firstName;
    
    
    @NotBlank(message="Field 'lastName' cannot be empty!")
    private java.lang.String lastName;
    
    private java.lang.Boolean loggedIn;
    
    private java.lang.String email;


    
    public java.lang.String getUsername() {
        return this.username;
    }

    public void setUsername(java.lang.String username) {
        this.username = username;
    }
    
    public java.lang.String getPassword() {
        return this.password;
    }

    public void setPassword(java.lang.String password) {
        this.password = password;
    }
    
    public java.lang.String getFirstName() {
        return this.firstName;
    }

    public void setFirstName(java.lang.String firstName) {
        this.firstName = firstName;
    }
    
    public java.lang.String getLastName() {
        return this.lastName;
    }

    public void setLastName(java.lang.String lastName) {
        this.lastName = lastName;
    }
    
    public java.lang.Boolean getLoggedIn() {
        return this.loggedIn;
    }

    public void setLoggedIn(java.lang.Boolean loggedIn) {
        this.loggedIn = loggedIn;
    }
    
    public java.lang.String getEmail() {
        return this.email;
    }

    public void setEmail(java.lang.String email) {
        this.email = email;
    }
    

    
    public java.lang.String getId(){
        return this.username;
    }
    
}